Rename TestClusterBasedJobCoordinatorRunner back to TestClusterBasedJobCoordinator Add TestClusterBasedJobCoordinatorRunner to test runClusterBasedJobCoordinator() in ClusterBasedJobCoordinatorRunner
